VGG19/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
joblib==1.4.2
keras==3.7.0
librosa==0.10.2.post1
matplotlib==3.9.3
numpy==2.0.2
pillow==11.0.0
scikit-learn==1.5.2
tensorflow==2.18.0
requests==2.32.3